requirements.txt
streamlit>=1.28.0
google-generativeai>=0.8.0
pymupdf>=1.23.0
Pillow>=10.0.0
pandas>=2.0.0
requests>=2.28.0
beautifulsoup4>=4.12.0
openai>=1.0.0
# PII 마스킹 (이미지 내 주민번호 뒷자리 검은색 처리)
easyocr>=1.7.0
opencv-python-headless>=4.8.0
numpy>=1.24.0